backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
Bottleneck==1.6.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.0
filelock==3.20.2
filetype==1.2.0
fsspec==2025.12.0
google==3.0.0
google-auth==2.46.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-core==1.2.6
langchain-google-genai==4.1.3
langchain-google-vertexai==3.2.1
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numexpr==2.14.1
numpy==2.4.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
pillow==12.1.0
pinecone==8.0.0
pinecone-client==6.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
proto-plus==1.27.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyMuPDF==1.26.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
validators==0.35.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0

# Authentication & Database
motor
pymongo
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.4.0
itsdangerous==2.2.0

# Cloud storage
cloudinary


